+++ src/ScottPlot5/ScottPlot5 Demos/ScottPlot5 WinForms Demo/Demos/DraggableAxisSpans.cs
@@ -1,4 +1,5 @@
 ﻿using ScottPlot;
+using ScottPlot.Plottables;
 
 namespace WinForms_Demo.Demos;
 
@@ -9,7 +10,7 @@ public partial class DraggableAxisSpans : Form, IDemoWindow
     public string Description => "Demonstrates how to create a mouse-interactive " +
         "axis span that can be resized or dragged";
 
-    DraggedSpan? Dragging = null;
+    AxisSpanUnderMouse? SpanBeingDragged = null;
 
     public DraggableAxisSpans()
     {
@@ -18,8 +19,15 @@ public partial class DraggableAxisSpans : Form, IDemoWindow
         // place axis spans on the plot
         formsPlot1.Plot.Add.Signal(Generate.Sin());
         formsPlot1.Plot.Add.Signal(Generate.Cos());
-        formsPlot1.Plot.Add.VerticalSpan(.23, .78);
-        formsPlot1.Plot.Add.HorizontalSpan(23, 42);
+
+        var vs = formsPlot1.Plot.Add.VerticalSpan(.23, .78);
+        vs.IsDraggable = true;
+        vs.IsResizable = true;
+
+        var hs = formsPlot1.Plot.Add.HorizontalSpan(23, 42);
+        hs.IsDraggable = true;
+        hs.IsResizable = true;
+
         formsPlot1.Refresh();
 
         // use events for custom mouse interactivity
@@ -30,39 +38,52 @@ public partial class DraggableAxisSpans : Form, IDemoWindow
 
     private void FormsPlot1_MouseDown(object? sender, MouseEventArgs e)
     {
-        Dragging = DraggedSpan.GetThingUnderMouse(formsPlot1, e.X, e.Y);
-
-        // disable pan/zoom while a plottable is being dragged
-        if (Dragging is not null)
+        var thingUnderMouse = GetSpanUnderMouse(e.X, e.Y);
+        if (thingUnderMouse is not null)
         {
-            formsPlot1.Interaction.Disable();
+            SpanBeingDragged = thingUnderMouse;
+            formsPlot1.Interaction.Disable(); // disable panning while dragging
         }
     }
 
     private void FormsPlot1_MouseUp(object? sender, MouseEventArgs e)
     {
-        Dragging = null;
-        formsPlot1.Interaction.Enable();
+        SpanBeingDragged = null;
+        formsPlot1.Interaction.Enable(); // enable panning
         formsPlot1.Refresh();
     }
 
     private void FormsPlot1_MouseMove(object? sender, MouseEventArgs e)
     {
-        if (Dragging is not null)
+        if (SpanBeingDragged is not null)
         {
-            // currently draggins something so update it
-            Coordinates cs = formsPlot1.Plot.GetCoordinates(e.X, e.Y);
-            Dragging.Update(cs);
+            // currently dragging something so update it
+            Coordinates mouseNow = formsPlot1.Plot.GetCoordinates(e.X, e.Y);
+            SpanBeingDragged.DragTo(mouseNow);
             formsPlot1.Refresh();
         }
         else
         {
-            // not dragging anything
-            DraggedSpan? spanUnderMouse = DraggedSpan.GetThingUnderMouse(formsPlot1, e.X, e.Y);
+            // not dragging anything so just set the cursor based on what's under the mouse
+            var spanUnderMouse = GetSpanUnderMouse(e.X, e.Y);
             if (spanUnderMouse is null) Cursor = Cursors.Default;
-            else if (spanUnderMouse.IsVerticalEdge) Cursor = Cursors.SizeWE;
-            else if (spanUnderMouse.IsHorizontalEdge) Cursor = Cursors.SizeNS;
-            else Cursor = Cursors.SizeAll;
+            else if (spanUnderMouse.IsResizingHorizontally) Cursor = Cursors.SizeWE;
+            else if (spanUnderMouse.IsResizingVertically) Cursor = Cursors.SizeNS;
+            else if (spanUnderMouse.IsMoving) Cursor = Cursors.SizeAll;
         }
+    }
+
+    private AxisSpanUnderMouse? GetSpanUnderMouse(float x, float y)
+    {
+        CoordinateRect rect = formsPlot1.Plot.GetCoordinateRect(x, y, radius: 10);
+
+        foreach (AxisSpan span in formsPlot1.Plot.GetPlottables<AxisSpan>().Reverse())
+        {
+            AxisSpanUnderMouse? spanUnderMouse = span.UnderMouse(rect);
+            if (spanUnderMouse is not null)
+                return spanUnderMouse;
+        }
+
+        return null;
     }
 }