+++ src/ScottPlot5/ScottPlot5/Plottables/HorizontalSpan.cs
@@ -33,4 +33,49 @@ public class HorizontalSpan : AxisSpan, IPlottable
         PixelRect rect = new(horiz, vert);
         Render(rp, rect);
     }
+
+    public override AxisSpanUnderMouse? UnderMouse(CoordinateRect rect)
+    {
+        AxisSpanUnderMouse spanUnderMouse = new()
+        {
+            Span = this,
+            MouseStart = rect.Center,
+            OriginalRange = new CoordinateRange(X1, X2),
+        };
+
+        if (IsResizable && rect.ContainsX(X1))
+        {
+            spanUnderMouse.ResizeEdge1 = true;
+            return spanUnderMouse;
+        }
+        else if (IsResizable && rect.ContainsX(X2))
+        {
+            spanUnderMouse.ResizeEdge2 = true;
+            return spanUnderMouse;
+        }
+        else if (IsDraggable && rect.XRange.Intersects(XRange))
+        {
+            return spanUnderMouse;
+        }
+
+        return null;
+    }
+
+    public override void DragTo(AxisSpanUnderMouse spanUnderMouse, Coordinates mouseNow)
+    {
+        if (spanUnderMouse.ResizeEdge1)
+        {
+            X1 = mouseNow.X;
+        }
+        else if (spanUnderMouse.ResizeEdge2)
+        {
+            X2 = mouseNow.X;
+        }
+        else
+        {
+            double dX = spanUnderMouse.MouseStart.X - mouseNow.X;
+            X1 = spanUnderMouse.OriginalRange.Min - dX;
+            X2 = spanUnderMouse.OriginalRange.Max - dX;
+        }
+    }
 }