+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 * Plot: Added `Plot.GetPixel()` overload for improved support on multi-axis plots (#3306) _Thanks @MCF_
 * Axis Lines: Label background color may be distinct from line color (#3309) _Thanks @PhoenixChenLu_
 * Axis Spans: New `Plot.Add.HorizontalSpan()` and `Plot.Add.VerticalSpan()` methods for shading axis ranges (#3307) _Thanks @erikjl_
-* Demo: Added an example window demonstrating draggable axis spans (#3307) _Thanks @erikjl_
+* Axis Spans: Added support for draggable axis spans (#3307) _Thanks @erikjl_
 
 ## ScottPlot 5.0.20
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-01-21_